% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion_utils.R
\name{random_D}
\alias{random_D}
\title{Generate random variate from density function for D.}
\usage{
random_D(sigmaD2 = 1, D0 = 0, time = 10^3, Dstar = 100,
  num_sample = 100)
}
\arguments{
\item{sig2}{is the diffusion rate parameters. Equation 4 Thompson et al. 2016.}
}
\description{
See equations 5 and 6 in Thompson et al. 2016. Genetics
}
