% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_utils.R
\name{run_stochastic_sim}
\alias{run_stochastic_sim}
\title{Simulate two paralogs evolving by compensatory drift.}
\usage{
run_stochastic_sim(p1_start = 5, p2_start = 5, theta = 10,
  sim_time = 10^4, mu = 10^-2, N = 10^2, sigma_m = 1, omega = 2,
  trace = FALSE, sample_rate = 10000, num_sample = 1)
}
\arguments{
\item{trace}{will plot out the simulations sampled every sample_rate mutations}
}
\description{
Simulate two paralogs evolving by compensatory drift.
}
